Remove incorrect comment
The comment says that mypy gives a "Missing return statement" but it
links to an issue on the mypy bug tracker about the same error message
reported after exhaustive enum matching. This is a completely different
situation, here the "Missing return statement" is warranted: this part
of the code is reachable so an explicit `return None` is wanted for this
function that returns an Optional.